Move template modules into templates/ subdirs
Refactor: relocate naming-template-engine to src/lib/templates/ and move naming-template, retention-policy, and schedule-preset services to src/services/templates/. Update imports across the app, components, runner step, and unit tests to the new paths. Fix: getNamingTemplate now returns the found template (previously only threw on not-found); the corresponding unit test now asserts the returned value. Tests: expand naming-template-engine tests with edge cases (empty pattern, plain-text passthrough, date-token-in-job-name, and timezone day-boundary shift). Update changelog to document the reorganization and test changes.
