+++ src/app/actions/templates.ts
@@ -4,9 +4,9 @@ import { auth } from "@/lib/auth";
 import { headers } from "next/headers";
 import { checkPermission, getUserPermissions } from "@/lib/auth/access-control";
 import { PERMISSIONS } from "@/lib/auth/permissions";
-import * as retentionPolicyService from "@/services/retention-policy-service";
-import * as namingTemplateService from "@/services/naming-template-service";
-import * as schedulePresetService from "@/services/schedule-preset-service";
+import * as retentionPolicyService from "@/services/templates/retention-policy-service";
+import * as namingTemplateService from "@/services/templates/naming-template-service";
+import * as schedulePresetService from "@/services/templates/schedule-preset-service";
 import { revalidatePath } from "next/cache";
 import { scheduler } from "@/lib/server/scheduler";
 import { logger } from "@/lib/logging/logger";
